Ability to build many projects via shared ffbt and shared node_modules
