+++ .npmrc
@@ -1,1 +1,2 @@
 save-prefix=
+package-lock=false
